+++ src/main/java/dev/nullkeeperdev/nulltweaks/feature/freecam/FreecamFeature.java
@@ -3,7 +3,6 @@ package dev.nullkeeperdev.nulltweaks.feature.freecam;
 import com.google.gson.JsonObject;
 import com.mojang.blaze3d.platform.InputConstants;
 import dev.isxander.yacl3.api.ButtonOption;
-import dev.isxander.yacl3.api.ConfigCategory;
 import dev.isxander.yacl3.api.Option;
 import dev.isxander.yacl3.api.OptionDescription;
 import dev.isxander.yacl3.api.OptionGroup;
@@ -125,14 +124,10 @@ public final class FreecamFeature extends Feature {
     }
 
     @Override
-    public void buildConfig(ConfigCategory.Builder builder) {
-        builder.group(OptionGroup.createBuilder()
-                .name(Component.literal("Freecam"))
-                .collapsed(false)
-                .option(speedOption())
+    public void buildConfig(OptionGroup.Builder builder) {
+        builder.option(speedOption())
                 .option(showHandOption())
-                .option(keybindButton())
-                .build());
+                .option(keybindButton());
     }
 
     @Override